Initial overhaul of menu UI and new print_at method
